+++ GoRogue.UnitTests/FOVTests.cs
@@ -124,5 +124,33 @@ namespace GoRogue.UnitTests
             foreach (var pos in fov.DoubleResultView.Positions())
                 Assert.Equal(0.0, fov.DoubleResultView[pos]);
         }
+
+        [Theory]
+        [MemberDataEnumerable(nameof(Radii))]
+        public void MultipleFOVDistanceOverlap(Radius shape)
+        {
+            var fov = new RecursiveShadowcastingFOV(_losMap);
+            var decay = 1.0 / (_radius + 1);
+            Distance dist = shape;
+
+            // Calculate two overlapping FOVs
+            var point2 = _center + 5;
+            fov.Calculate(_center, _radius, shape);
+            fov.CalculateAppend(point2, _radius, shape);
+
+            // Print result, for reference
+            _output.WriteLine("Resulting FOV (For Reference):");
+            _output.WriteLine(fov.ToString());
+
+            // Make sure the distance portion of the FOV always reflects the closest source (to ensure overlapping
+            // radii were handled properly)
+            foreach (var pos in fov.DoubleResultView.Positions())
+            {
+                double minDist = Math.Min(dist.Calculate(pos, _center), dist.Calculate(pos, point2));
+
+                double expectedVal = minDist > _radius ? 0 : 1.0 - (decay * minDist);
+                Assert.Equal(expectedVal, fov.DoubleResultView[pos]);
+            }
+        }
     }
 }